Change link definition syntax
Links are now to be indented by a single tab, and they are to be defined
_below_ their usage. Links can still be defined twice, the definition
immediately below the usage shall be used.

Input from stdin is not supported, as the program needs to go forward to
find the definition each time it encounters a link, and then come back.
(This behaviour may cause performance to be hit ?)