Allow links, HTML tags, etc. in the subtitle text
This elegant solution was hiding just under my nose the whole time!
Why the hell didn't I think of it earlier?

NOTE: If you are confused why this works, look at the raw blog layout.
Specifically, the subtitle text ends with the same "---\n" that the blog
text ends with. So, htmlize() works for both!