HUGE BUGFIX (misused static inside functions)
The effect of `static` on a variable is different when used inside
functions. It doesn't mean that the variable will be limited to the
current function (like `static` in file-level declarations mean that the
declared variable shall be visible to the current file only), because
that is the default behaviour _without_ `static` (since the function
statements are enclosed within {}, which forms a scope, and variables
are always scope-local by default). Rather, it works something like a
function-specific counter.

For eg, if a function contained the following two statements -

	static long long int counter = 0;
	counter++;

That would cause the value of counter to *not* reset when the current
function scope exits. Try it out yourself -

	#include <stdio.h>
	int main(void) { for (int i=0; i<10; i++) func(); }
	void func(void) {
		static int count = 0;
		putc('0' + count++);
		putc('\n');
	}